extract renderer-util.openMatch into match-details module
